Re-organize gui folder structure
