+++ gui/build_ui.bat
@@ -1,3 +1,3 @@
 pyside6-uic randogui.ui > ui_randogui.py
-pyside6-uic tricks_dialog.ui > ui_tricks_dialog.py
-pyside6-uic custom_theme_dialog.ui > ui_custom_theme_dialog.py
+pyside6-uic dialogs/tricks/tricks_dialog.ui > dialogs/tricks/ui_tricks_dialog.py
+pyside6-uic dialogs/custom_theme/custom_theme_dialog.ui > dialogs/custom_theme/ui_custom_theme_dialog.py
