+++ gui/dialogs/tricks/tricks_dialog.ui
