+++ gui/dialogs/tricks/ui_tricks_dialog.py
