+++ gui/themes/default_theme_info.yaml
